putnbr gere long

+++ ft_putnbr.c
@@ -6,17 +6,17 @@
 /*   By: mmouhssi <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/12/01 18:37:58 by mmouhssi          #+#    #+#             */
-/*   Updated: 2015/12/09 17:35:30 by mmouhssi         ###   ########.fr       */
+/*   Updated: 2016/04/08 00:51:00 by mmouhssi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_putnbr(int n)
+void	ft_putnbr(long n)
 {
 	long nbr;
 
-	nbr = (long)n;
+	nbr = n;
 	if (nbr < 0)
 	{
 		ft_putchar('-');
